Added Topic Validation

+++ server/resources/controller/topicController.js
@@ -1,8 +1,10 @@
 const Topic = require('../models/Topic')
-
+const {
+    addTopicValidation,
+    topicValidation,
+} = require("../utils/validation/topicValidation");
 
 const getAllTopics = async ( req, res ) => {
-
     try{
         let result = await Topic.find().exec()
         res.status(200).json({error: false, data: result})
@@ -27,7 +29,9 @@ const getTopicsSummary = async ( req, res ) => {
 
 const getTopicById = async ( req, res ) => {
     const { id } = req.params
-
+    if (id === undefined) {
+        return res.status(400).json({ error: true, message: "Pass Topic ID" });
+    }
     try{
         let result = await Topic.findById(id).exec()
         if(!result){
@@ -42,7 +46,13 @@ const getTopicById = async ( req, res ) => {
 
 const getTopicByName = async ( req, res ) => {
     let { name } = req.params
-    
+    const { error } = topicValidation(req.params)
+    if (error) {
+        return res.status(400).json({
+            error: true,
+            reason: error.details[0].message,
+        });
+    }
     try{
         name = name.toUpperCase()
         let result = await Topic.findOne({name}).exec()
@@ -58,7 +68,13 @@ const getTopicByName = async ( req, res ) => {
 
 const addTopic = async ( req, res ) => {
     const { name, icon } = req.body  
-
+    const { error } = addTopicValidation(req.body)
+    if (error) {
+        return res.status(400).json({
+            error: true,
+            reason: error.details[0].message,
+        });
+    }
     try{
         let temp = await Topic.findOne({name}).exec()
         if(!temp){
@@ -73,10 +89,14 @@ const addTopic = async ( req, res ) => {
         res.status(400).json({error: true, message: err})
     }
 }
- 
+
 const deleteTopic = async ( req, res ) => {
     const { id } = req.params
-
+    if (id === undefined) {
+        return res
+        .status(400)
+        .json({ error: true, message: "Pass Topic ID" });
+    }
     try{
         let result = await Topic.deleteOne({_id: id})
         if(result.deletedCount === 0){
@@ -92,7 +112,13 @@ const deleteTopic = async ( req, res ) => {
 const editTopic = async ( req, res ) => {
     const { id } = req.params
     const data = req.body
-
+    console.log(Object.keys(data));
+    if (id === undefined) {
+        return res.status(400).json({ error: true, message: "Pass Topic ID" });
+    }
+    if (!Object.keys(data).length) {
+        return res.status(400).json({ error: true, message: "Pass Data" });
+    }
     try{
         let result = await Topic.updateOne({_id: id}, data)
         res.status(200).json({error: false, message: "The topic has been updated"})
@@ -105,7 +131,13 @@ const editTopic = async ( req, res ) => {
 const replaceTopic = async ( req, res ) => {
     const { id } = req.params
     const data = req.body
-
+    const { error } = addTopicValidation(req.body);
+    if (error) {
+      return res.status(400).json({
+        error: true,
+        reason: error.details[0].message,
+      });
+    }
     try{
         let result = await Topic.replaceOne({_id: id}, data)
         res.status(200).json({error: false, message: "The topic has been replaced"})
